requirements.txt
python-telegram-bot==22.5
easyocr
opencv-python
google-api-python-client
google-auth
google-auth-oauthlib